Create PurviewKickstart.ps1
This is an uber automation script which intends to create RG and Purview Account along with the ADF provisioning when a customer tries a demo trial of purview using powershell. The benefit of this script will be the customer doesn't need to switch between portal and script back and forth, just running this script should set up everything needed for Purview from scratch. Refer : https://docs.microsoft.com/en-us/azure/purview/tutorial-scan-data